Move XmlFollowSchema from Cfdi33 to common Xml namespace
